Ready for release build/release
 - Add SwiftRadioCoverSnapshots target and scheme
 - Save current changes to the Fastfile
 - Update SnapshotHelper to current

Release builds properly. Snapshot creation has been updated to take
snapshots of a track without a cover, followed by a new snapshot target
that waits until it sees an actual cover for the "play" and "share"
snapshots.

Added a frameit target to the Fastfile and verified it works. Ready to
upload screenshots and push the first release when I get approval from
the testers.